Add files via upload
This isn't merely a resume; it's the genesis of TMinus36, a brand encapsulating decades of innovation. Since 1995, I, Timothy Harris, have operated under the moniker TMinus36, a testament to a lifetime dedicated to pushing the boundaries of IT and related technologies. My journey culminated in the founding of Harris Secured Networks (HSN) in 2011. Together, these entities represent more than just a professional history—they embody the foundational elements of a technological legacy poised to transcend generations.